backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.5
fal-client
